requirements.txt
streamlit
agno
python-dotenv
google-genai

